+++ src/models/Product.ts
@@ -46,7 +46,7 @@ export class Product extends BaseDeleteEntity {
   @Column({ type: 'boolean', default: false })
   enableUrlRedirect: boolean;
 
-  @Column({ type: 'bigint', default: 0 })
+  @Column({ type: 'float', default: 0 })
   price: number;
 
   @Column({ type: 'text', nullable: true })
